// MotoPronto/src/app/models/usuario.ts
export interface Usuario {
    idUsuario?: number;
    nombreUsuario: string;
    apellidoP: string;
    apellidoM: string;
    telefono: string;
    correoElectronico: string;
    ciudad: string;
    codigoPostal: string;
    calleNumero: string;
    password: string;
    colonia: string;
    idRol: number;
    avatar: string;
  }
  